PWN-6355 - Resolve PR comments
