+++ app/src/main/java/org/p2p/wallet/moonpay/repository/sell/MoonpaySellFiatCurrency.kt
@@ -1,6 +1,6 @@
 package org.p2p.wallet.moonpay.repository.sell
 
-enum class MoonpaySellFiatCurrency(val symbol: String, val currencySymbol: String) {
+enum class MoonpaySellFiatCurrency(val symbol: String, val uiSymbol: String) {
     EUR("eur", "€"),
     USD("usd", "$"),
     GBP("gbp", "GBP");
