+++ app/src/main/java/org/p2p/wallet/sell/ui/payload/SellPayloadContract.kt
@@ -1,11 +1,29 @@
 package org.p2p.wallet.sell.ui.payload
 
+import androidx.annotation.ColorRes
 import org.p2p.wallet.common.mvp.MvpPresenter
 import org.p2p.wallet.common.mvp.MvpView
 import org.p2p.wallet.utils.Base58String
 import java.math.BigDecimal
 
 interface SellPayloadContract {
+
+    data class CashOutButtonState(
+        val isEnabled: Boolean,
+        @ColorRes val backgroundColor: Int,
+        @ColorRes val textColor: Int,
+        val text: String
+    )
+
+    data class ViewState(
+        val quoteAmount: String,
+        val fee: String,
+        val fiat: String,
+        val solToSell: String,
+        val tokenSymbol: String,
+        val fiatSymbol: String,
+        val userBalance: String
+    )
 
     interface View : MvpView {
         fun showLoading(isVisible: Boolean)
